+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release 3.0.7 (2021-10-31)
 - FIX: Issue when holding Alt for wall snapping whilc Ctrl is pressed/locked.
+- DEV: Migrated to a Webpack/Gulp build chain.
 
 ## Release 3.0.6 (2021-10-23)
 - FIX: Capture Canvas issue with scene with no background image.